+++ packages/eslint-plugin-browser-security/src/index.ts
@@ -181,7 +181,7 @@ withCanonicalDocsUrls('plugin-browser-security', rules);
 export const plugin: TSESLint.FlatConfig.Plugin = {
   meta: {
     name: 'eslint-plugin-browser-security',
-    version: '2.1.8',
+    version: '2.1.9',
   },
   rules,
 } satisfies TSESLint.FlatConfig.Plugin;